Update Code Standard Fibonacci

+++ Code/StandardFibonacci/StandardFibonacci/Fibonacci.cs
@@ -8,11 +8,11 @@ namespace StandardFibonacci
 {
     class Fibonacci
     {
-        public int get(int n)
+        public double get(int n)
         {
-            if (n == 0) return 0;
-            else if (n == 1) return 1;
-            else return get(n - 1) + get(n - 2);
+            if (n == 0)         return 0;
+            else if (n == 1)    return 1;
+            else                return get(n - 1) + get(n - 2);
         }
     }
 }
